Run everything through prettier-ruby and rubocop

+++ app/services/cms_content_loaders/base.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 class CmsContentLoaders::Base < CivilService::Service
   validate :ensure_no_conflicting_content
 
@@ -18,31 +19,27 @@ class CmsContentLoaders::Base < CivilService::Service
   end
 
   def load_content(&block)
-    content_set.all_template_paths_with_names(subdir).each do |path, name|
-      next if content_names.present? && !content_names.include?(name)
+    content_set
+      .all_template_paths_with_names(subdir)
+      .each do |path, name|
+        next if content_names.present? && content_names.exclude?(name)
 
-      content, attrs = content_set.template_content(path)
-      model = load_item(name, content, attrs)
-      next if model == :skip
+        content, attrs = content_set.template_content(path)
+        model = load_item(name, content, attrs)
+        next if model == :skip
 
-      block.call(model) if block
-    end
+        yield(model) if block
+      end
   end
 
   def load_item(name, content, attrs)
     if existing_content_identifiers.include?(name)
       return :skip if conflict_policy == :skip
 
-      if conflict_policy == :overwrite
-        convention_association.where(identifier_attribute => name).destroy_all
-      end
+      convention_association.where(identifier_attribute => name).destroy_all if conflict_policy == :overwrite
     end
 
-    convention_association.create!(
-      identifier_attribute => name,
-      content_attribute => content,
-      **attrs
-    )
+    convention_association.create!(identifier_attribute => name, content_attribute => content, **attrs)
   end
 
   def subdir
@@ -50,18 +47,15 @@ class CmsContentLoaders::Base < CivilService::Service
   end
 
   def identifier_attribute
-    raise NotImplementedError,
-      'CmsContentLoaders::Base subclasses must implement #identifier_attribute'
+    raise NotImplementedError, 'CmsContentLoaders::Base subclasses must implement #identifier_attribute'
   end
 
   def content_attribute
-    raise NotImplementedError,
-      'CmsContentLoaders::Base subclasses must implement #content_attribute'
+    raise NotImplementedError, 'CmsContentLoaders::Base subclasses must implement #content_attribute'
   end
 
   def convention_association
-    raise NotImplementedError,
-      'CmsContentLoaders::Base subclasses must implement #convention_associationd'
+    raise NotImplementedError, 'CmsContentLoaders::Base subclasses must implement #convention_associationd'
   end
 
   def taken_special_identifiers
@@ -75,15 +69,19 @@ class CmsContentLoaders::Base < CivilService::Service
   def ensure_no_conflicting_content
     return unless conflict_policy == :error
 
-    content_set.all_template_names(subdir).each do |name|
-      if taken_special_identifiers[name]
-        errors.add(:base, "#{convention.name} already has a #{taken_special_identifiers[name]}")
-      end
+    content_set
+      .all_template_names(subdir)
+      .each do |name|
+        if taken_special_identifiers[name]
+          errors.add(:base, "#{convention.name} already has a #{taken_special_identifiers[name]}")
+        end
 
-      if existing_content_identifiers.include?(name)
-        errors.add(:base, "A #{subdir.singularize} named #{name} already exists in \
-#{convention.name}")
+        next unless existing_content_identifiers.include?(name)
+                  errors.add(
+                    :base,
+                    "A #{subdir.singularize} named #{name} already exists in \
+        #{convention.name}"
+                  )
       end
-    end
   end
 end
